Move cacheProvider fn out of ui-common
