+++ ui-common/package.json
@@ -4,11 +4,7 @@
   "scripts": {
     "tsc:check": "tsc"
   },
-  "dependencies": {
-    "eth-rpc-cache": "2.0.0"
-  },
   "devDependencies": {
-    "@ethersproject/providers": "5.7.2",
     "next": "14.2.16",
     "react": "18.2.0",
     "react-dom": "18.2.0"
